Remove an unnecessary `/api/groups` call on page load
On startup an initial `FRAME_CONNECTED` event is fired when the main
frame connects to the sidebar. This triggered an extra `/api/groups`
fetch. Avoid this by recording the main frame URI used in the last
groups fetch in  `load()`. The existing test in the `FRAME_CONNECTED`
handler then skips the `load()` call if this URI has not changed.