Add an option to API request methods to include metadata in the response
Add an option to return metadata in the response as well as just the
JSON response from the server. This will initially be used to determine
the access token that was used to make the request and therefore whether
it was authenticated or not. In future it could also return information
from HTTP headers or the status code if useful.