Hide the "Public" group from the client under certain conditions
Implement a business requirement that the "Public" group should be
hidden from the groups list if all these conditions are met:

 - The user is logged-out
 - The current document has groups associated with it
 - The user has not visited a direct-link to an annotation in "Public"

---

This requirement introduces a complexity in that deciding which group to
focus and fetch annotations for after initially fetching groups now
depends on the current profile. However the profile and groups are
fetched concurrently to reduce the number of API round-trips before we
can show annotations in the client.

Fortunately we don't need to know the exact userid to know which group
to focus, only whether the list of groups was fetched as a logged-in
user or not. This can be determined from the access token that was sent
with the `/api/groups` request.