Create qa deployment of the client
Add a new build stage to Jenkins CI builds of the master branch which
deploys a "QA" build to S3. This QA build:

 - Has a pre-release version number of the form `X.Y.Z-{commit hash}`
 - Has a stable URL at https://cdn.hypothes.is/hypothesis@qa
   (as opposed to https://cdn.hypothes.is/hypothesis for the prod build)
 - Loads the sidebar app from qa-h (https://qa.hypothes.is/app.html)
 - Does not update the changelog or create a Git tag or GitHub release

This build is currently not deployed to npm at all, but we might need to
do that in future if we were to create a QA deployment of the browser
extension with it.

---

 - Add a QA deploy step to the Jenkins build.
 - Skip GitHub release creation and changelog updates for pre-release
   versions used for QA deployments.
 - Remove the preversion.sh script, which is no longer needed since the
   builds always run in Jenkins and not on a developer's system.
 - Support specifying the npm tag to check in `wait-for-npm-release.sh`,
   though this is not used in the final version of this commit.