Rewrite used parts of stream-filter.coffee in JS
Port stream-filter.coffee to JS and add tests, but only for the
functionality that we actually still use in the app. Following the
removal of support for real-time updates in the stream, the websocket is
only configured to filter by URI, annotation ID or reply IDs
("references") in the sidebar and standalone annotation viewer.