Use alternate method to get document fingerprint
`PDFViewerApplication.documentFingerprint` was removed in recent
versions of PDF.js [1]. This issue causes loading the bookmarklet into
Firefox 63+ to fail because the client thinks the PDF is still loading.

`PDFViewerApplication.pdfDocument.fingerprint` is the replacement
mentioned in the commit that removed the property and it also appears to
be more clearly a part of PDF.js' documented API.

This change is backwards-compatible with the version of PDF.js that
currently ships with Via.

[1] https://github.com/mozilla/pdf.js/commit/e522b2d87e8eb3d4eb8a391a098a0f6f157e8c12