Enable third-party authorities to globally disable annotation share links
In the context of the LMS app, the client is embedded inside a Via iframe, which
is inside an LMS app iframe, which is inside Canvas. There is currently no
mechanism in place to propagate the focused annotation down the frame hierarchy.
As a result, share links show up but do not work.

As a simple way to resolve that for an initial version of the LMS app, provide
it with a way to globally disable annotation share links.

In future we may move this responsibility to the "h" service, so that the
availability (or not) of share links is correctly reflected in h API responses.

Part of https://github.com/hypothesis/product-backlog/issues/843