Skip over `<link>` and `<meta>` tags with invalid URLs
Websites may put unparseable URLs in `<link>` and `<meta>` tags [1]. We
could choose to preserve the URL string as-is (which
`HTMLLinkElement#href` does for example) or not include it at all. I've
opted to skip over such URLs to make life easier for code consuming the
parsed links as it can rely on them being parseable as URLs.

[1] https://github.com/hypothesis/product-backlog/issues/742