Refactor Makefile
Refactor Makefile to be the same as for our other projects:

    $ make help
    make help              Show this help message
    make dev               Run the app in the development server
    make lint              Run the code linter(s) and print any warnings
    make test              Run the unit tests
    make docs              Build docs website and serve it locally
    make checkdocs         Crash if building the docs website fails
    make clean             Delete development artefacts (cached files,
                           dependencies, etc)

Other make commands that our other projects have but the client doesn't
yet:

`make shell` -- Not relevant to the client? (Unless we can get a useful
JavaScript shell in the terminal.)

`make sql` -- Not relevant.

`make format` and `checkformatting` -- We can't use Black code
formatting for the client of course but maybe we could use Prettier or
something else. (Our Python apps could also be extended to have `make
format` format the JavaScript code as well.)

`make coverage` and `codecov` -- Print the unit test coverage report to
the terminal, and post it to codecov. These should be doable?