Remove CHANGELOG from source tree
The client deployment process currently involves a commit to master to
update the CHANGELOG and version in package.json. This creates a problem
when trying to deploy a build that is not the latest release of master.
It also creates other failure modes that are not really necessary.

As a first step to resolving this problem, remove the changelog in the source
tree. The notes attached to the GitHub release will suffice, together
with any handwritten updates that developers post in #changelog in
Slack.

 - Remove CHANGELOG.md and scripts to generate it from source tree
 - Remove changelog references from build process
 - Change GitHub release script to generate the change list directly
   instead of reading it from CHANGELOG.md
 - Change the client's dev server to use the Code of Conduct instead of
   the changelog as a text source for the test page