Stop updating package.json version field during deployments from Jenkins
Creating a commit which updates the version number as part of
the release process prevents deployments of the master branch which are
not the latest commit, which is a surprising limitation compared to our
release process for h etc.

This commit resolves the issue by removing the version-bumping commit.
Instead package.json in the source tree contains a fixed, unchanging
version number which is set to the real version just before the package
is published to npm. The release version is obtained by fetching the tag
with the highest version number and incrementing the major part.