Refactor polyfill bundling and loading
Refactor polyfill bundling and loading to load only the polyfills that
are required for the current browser, and make it easier to load new
polyfills conditionally in future.

The polyfills are partioned into sets. For each set there is a separate
bundle, with an entry point in src/shared/polyfills/{set}.js. There is
an index file in src/shared/polyfills/index.js which defines functions
to test which sets are needed by the current browser.

This change prevents loading of ~246KB of minified JS (not gzipped) in modern
browsers which don't need any of the polyfills.