Enable async-await support
To facilitate writing more readable code that deals with promises,
especially when multi-step promise chains are involved, enable support for
async-await.

I have opted to use the babel-plugin-transform-async-to-promises plugin
[1] for this because it generates code which is much shorter and easier to
read/debug than the "default" approach of transform async functions first to
generators and then using regenerator-runtime.

In working on this I discovered that ~90% of client users have browsers
with native async-await support [2], so it may not be too long before we
should consider building an untranspiled version of the code and loading
that in most browsers instead.

[1] https://github.com/rpetrich/babel-plugin-transform-async-to-promises
[2] https://hypothes-is.slack.com/archives/C1M8NH76X/p1552575007000100?thread_ts=1552499427.025200&cid=C1M8NH76X