Port h-tooltip directive to React
There are various ways that an Angular directive could be translated to
React. The approach taken here is to create a hook, `useTooltip`, which
creates a function that accepts a DOM element and shows a tooltip when
that element is hovered. This function can be passed as the
`ref` prop for an element and React will call it when the DOM node is
created, changed or removed.

This was informed by the approach in
https://reactjs.org/docs/hooks-faq.html#how-can-i-measure-a-dom-node

One disadvantage is that because the same hooks must be called in the
same order on every render, `useTooltip` must be called if it _might_ be
used, whether or not the element with the tooltip is actually rendered.