Remove loading based on conditional feature flag
Since the feature flag isn't available until the profile api request
completes, and the profile api request is run asynchronously with the
group api requests, it doesn't make sense to load conditionaly. Instead,
call all the group api requests and filter the groups for the old groups
menu later. In this case, filtering is done in the groups.all
conditionally based on the feature flag.