Fix npm dependencies make target not re-installing if yarn lockfile changed
The make target used to ensure that npm dependencies are installed when
they change depended on `yarn.lock` but didn't actually run `yarn
install` if only that file, and not package.json, changed.

The target relied on a tool called check-dependencies which pre-dates
yarn. `yarn install` is fast if nothing changed, so just run it
unconditionally.