Fix initialization of visibleThreadsList
Previously the visibleThreadList state in thread-list did not get
calculated in onInit because the method that did so was subscribed to
after the 'changed' event was emitted. This issue was resolved by
calling the method to calculate the state explicitly in onInit and
only subscribing the event handler for changes that may occur after init.