Make group selectors more efficient
Ensure that selectors for getting the "My groups", "Currently viewing
groups" and "Featured groups" lists only recompute when the groups state
changes. Previously they would have recomputed when _any_ part of the state changed.

More generally, we should be using `createSelector` rather than
`memoize` whenever we have a place in the app that needs to extract a subset
of data from some larger structure and derive some data from that.