Convert timestamp component to React
A notable change is that the current time used to generate the relative
timestamp is now stored as an explicit piece of state, rather than
implicitly using the current time when calling `toFuzzyString`. This
provides a clean way to trigger a re-render of timestamp periodically:
an effect updates the "now" time, and that triggers a recalculation of
the timestamp.

The tests have also been revised to better capture the user-facing
behavior of the component.