Fix overlapping text in headings in annotation bodies
Parent elements of annotation cards set line heights to absolute pixel
values. Annotation card headings then inherit this and as a result,
their line height is too small for the font size.

We ought to reconsider the way line heights are set elsewhere in the
app, but resetting the line height in the `styled-text` mixin ensures
that annotation body styles are protected from whatever the parent
elements do with it.