Replace several uses of `console.warn` with `warnOnce`
This reduces warning spam in the event where the code path that triggers
the warning is invoked repeatedly and there is no value to the developer
in telling them about the issue more than once.

To make `warnOnce` an easier drop-in replacement for `console.warn`, it
has been modified to accept multiple arguments, all of which get
forwarded to `console.warn`.