Remove wrapping `<span>` around SVG icons
This makes it easier to get icons positioned correctly because there is one
less element in the hierarchy to deal with and any class passed to
`SvgIcon` is applied to the root element rendered by that component.

The implementation works by first using the browser's HTML parser (via
`innerHTML`) to parse the SVG markup and then converting the returned
DOM tree to a Preact element, which the `SvgIcon` component returns as
its render output. The implementation is not fully generic, it only does what
is necessary to handle SVG icons.