Instead of using search.query use store directly
Previously a search object was built inside the hypothesis-app
component and passed down to its child components. Now, just use
the store.getState().filterQuery directly. This avoids unnecessarily
passing things between components and a layer of indirection between
the search object and the value in the store. Rename searchQuery to
filterQuery.