Update comment to reflect that both an annotation AND a group can be direct-linked
The URL fragment syntax for direct links only supports either an
annotation or a group, the client's JSON configuration however currently allows
both an annotation and a group to be specified. If that happens, the
client goes to the group containing the direct-linked annotation, but
will still add the direct-linked group to the groups list.