Replace references to old groups menu styles in new component
This replaces some references to the styles for the old groups menu in
the new implementation. The menu label is now slightly darker as it
uses the same grey as the menu items. Otherwise there should be no
noticeable visual changes.

The CSS itself has been simplified compared to the previous
implementation.