Add a utility function to copy text to the clipboard
I evaluated popular npm packages for this purpose (eg. [1]) but encountered
issues where they were not transpiled for use in older browsers or
included workarounds for browsers we no longer support (eg. IE 9).

If we do decide to use an existing package in future, we can keep this
utility module as a wrapper around it.

[1] https://github.com/sindresorhus/copy-text-to-clipboard