Add "Copy link" action to group submenu
Add an action to provide a convenient way to copy the sharing link
directly from the groups menu.

Depending on whether the group is private or open and whether the
receiving user is a member or not, the link may present the user with a
Join page or the activity page for the group. To simplify the client logic, the
text just says "Copy link".