Separate hover areas for main action and submenu toggle
For menu items that have associated submenus, make it clearer that there
are two separate click zones by highlighting only the clickable area for
the main action when hovering that part of the item.

For menu items without submenus, there should be no visual change.

The implementation adds a `.menu-item__action` element that wraps the
content of the menu item excluding the submenu toggle.

The CSS has been re-organized to group together all of the rules that
affect a specific element. This makes it easier to verify that there are
no states which visually conflict for a particular element.