Always use main URL from main frame when searching for groups
Previously the groups service used the first HTTP(S) URL returned by
`store.searchUris()` to get the URL used to search for associated groups. In
practice, the URIs from the main frame always appeared first in this
list, and the main URL from that frame was the first entry within that sublist.
However, the `store.searchUris` function doesn't provide any guarantee of this
and future changes could break this invariant.

This commit adds a `store.mainFrame()` function which returns the main
frame, the one with no `id`, and modifies the groups service to use the
main URI for that frame.

A change in behavior here is that if the top-level frame was a non-HTTPS
frame but embedded an HTTPS frame, then the group search would have used
the URL from the HTTPS frame. Now it will use the non-HTTPS URL from the
top-level frame.