Replace `MenuItem`'s `renderSubmenu` prop with `submenu` prop
Simplify the API of `MenuItem` a little by passing the content of the
submenu directly rather than a function to render the submenu. The
downside of this is that `MenuItem` cannot skip rendering the submenu
content when the submenu is not visible, but the implementation does not
do this anyway.