Implement client-side rate limiting per session
Previous experience with Sentry has found that there are various
scenarios that can cause a client to spam the server with a large number
of reports. Although Sentry has its own server-side quotas and rate
limiting, I think it will be useful to do client-side per-session rate
limiting as well. This also provides a place where we can hook in other
client-side filtering in future.