Whitelist Sentry error reports to only include errors from our code
To reduce noise from Sentry, filter events to exclude all errors which
cannot be attributed to our code based on the URLs in the error's stack
trace.

This will prevent us from receiving reports caused by eg. script tags
injected by browser extensions or custom browsers.

There is a downside to this: We will stop receiving error reports which
are caused by / occur in our code but cannot be attributed to it. This
can happen for example if a promise is rejected with a non-Error value
so no stack trace can be obtained.

For the moment, we're prioritising precision (% of error reports
actually caused by problems in our code) over recall (% of problems in
our code that are reported).