Set SIDEBAR_APP_URL to a sensible default value in dev
Set SIDEBAR_APP_URL to work with your local dev instance of h by default
in dev.

This is in line with the rest of our apps (h, lms, bouncer, via): when
run in dev they default to working with the dev instances of the others
app, not the production ones.

The idea is that, when setting up all the apps in a dev env, you just
have to install each app and run `make dev` and it will work correctly
with the other dev apps. Some apps also require a `make devdata` command
if they need sensitive DB contents and/or envvars.

Don't require developers to have to set lots and lots of envvars in each
individual project.

I also deleted some docs that instructed people to create authclients,
and set envvars, that are already automated in h.