Don't built the client a second time when running "npm publish"
Fix an issue with the CI build where both the `yarn version` and
`npm publish` commands would both build the client's assets.

Since the output of the build process can be affected by environment
variables, this could lead to unexpected results if env vars were set
when running `yarn version` but not when running `npm publish`.