Enable tracing for even more detailed debug log capturing
This enables very detailed logging of network activity, page execution
and script execution for 7 seconds after startup.

This is part of debugging frequent CI build failures in the client repo
on Jenkins. From the Karma logs I can see that all of the test files are
requested from the Karma server, but none of the log messages associated
with test execution are seen. This tracing should make it possible to
find out which JS scripts even execute.