Disable verbose Karma logging
This adds a lot of noise to the Jenkins log and we've gotten all the value we
can out of this for tracking down Jenkins CI build failures. What remains is a
period of time between Karma creating the `context.html` iframe where the test
code executes and the scripts in that iframe starting to actually execute.

During this time neither the Karma logs nor Chrome's verbose log provide any
insight into what is going on. Chrome tracing however does, so that's where the
investigation is continuing.