Enable throttling of concurrent builds
Jenkins CI builds of the client project are prone to failing when many
concurrent builds occur. This often happens when many Dependabot PRs are
submitted at once during a weekly dependency update. These builds
succeed most of the time when retried.

Limiting the number of concurrent builds per executor can be done globally,
but it can also be done on a per-project basis using the
throttle-concurrents [1] plugin. Assign a throttling category to client
builds so that this latter approach can be used.

[1] https://plugins.jenkins.io/throttle-concurrents