Remove "enabled" state from excerpt component
Simplify the `<excerpt>` component by removing the `enabled` property.
This property was used to avoid the markdown editor being truncated when
rendered inside the excerpt. However it is possible to achieve the same
result without this property by using `ng-if`.

This change will simplify porting the excerpt component to Preact and
also paves the way to splitting the existing `<markdown>` component into
two separate, simpler, components: One for the "view" mode and one for the "edit"
mode.