Allow function props to be invoked during a digest cycle
Allow function props passed to Preact components wrapped with
`wrapReactComponent` to be invoked when a digest cycle is already active
for some reason. In that case using `$scope.$apply` triggers an error.

This scenario is rare, as callbacks are usually invoked in response to
events that happen outside of Angular, but changes made by Angular code
can indirectly cause Preact components to invoke callbacks. For example,
in layout effects which run at the end of a `render` call.