Add a new implementation of text position => Range conversion
Add a new implementation of conversion from text positions (that is,
offsets within an element's `textContent`) to DOM `Range`s along with
test cases.

This addresses an issue with the existing implementation of `toRange` in
the `dom-anchor-text-position` package where conversion fails when the
text position includes the end of the element's text.

Even if/when the issue is addressed upstream, I think it would be useful
to retain these test cases to guard against future regressions.

See #1329